actualizacion 1.1, se cambia el campo monto total para que pueda leer el dato que se almacenara

+++ new_module/models/models.py
@@ -83,7 +83,7 @@ class pruebas(models.Model):
         default=0.19,
         readonly=True
          )
-    monto_Total = fields.Integer(
+    monto_Total = fields.Float(
         'Monto Total',
         default=montoNeto*montoIvaRecuperable
         )
